fix: move DiskPressureBanner to shared components to fix cross-domain lint
